\name{optgradients}
\docType{data}
\alias{optgradients}
\title{Optimal gradient directions for number of gradients between 6 and 162}
\description{Optimal gradient directions minimizing symmetrized 
Coulomb forces on the sphere following a proposal by Jones et al.
(1999). These directions define an optimal design in DWI for given
number of gradients. 
}
\usage{optgrad}
\format{a list with name optgrad and component ngrad-5 containing
a matrix with ngrad gradients as columns.
}
\keyword{datasets}
