\name{showFAColorScale}
\alias{showFAColorScale}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Writes an image with the colqFA colorscale to disk.
}
\description{
  Writes an image (type PNG) with the colqFA colorscale to disk.
}
\usage{
  showFAColorScale(filename = "FAcolorscale.png")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{filename}{ Name of file to write.}
}
\seealso{
  See Also \code{\link{colqFA}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ manip }
