\name{dti-package}
\alias{dti-package}
\alias{dti}
\docType{package}
\title{
   \packageTitle{dti}
}
\description{
\packageDescription{dti}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dti}
\packageIndices{dti}
}
\author{
\packageAuthor{dti}

Maintainer: \packageMaintainer{dti}
}
\references{S. Mohammadi, K. Tabelow, L. Ruthotto, Th. Feiweier, J. Polzehl,
            and N. Weiskopf, \emph{High-resolution diffusion kurtosis imaging
            at 3T enabled by advanced post-processing}, 8 (2015), 427.

            S. Becker, K. Tabelow, S. Mohammadi, N. Weiskopf, and J. Polzehl,
            \emph{Adaptive smoothing of multi-shell diffusion weighted magnetic
            resonance data by msPOAS}, NeuroImage 95 (2014), pp. 90-105.

            S. Becker, K. Tabelow, H.U. Voss, A. Anwander, R.M. Heidemann and
            J. Polzehl, \emph{Position-orientation adaptive smoothing of
            diffusion weighted magnetic resonance data (POAS)}, Medical Image
            Analysis, 16 (2012), pp. 1142-1155.

            J. Polzehl and K. Tabelow,
            \emph{Beyond the diffusion tensor model: The package \pkg{dti}},
            Journal of Statistical Software, 44 no. 12 (2011) pp. 1-26.

            K. Tabelow, H.U. Voss and J. Polzehl,
            \emph{Modeling the orientation distribution function by
            mixtures of angular central Gaussian distributions},
            Journal of Neuroscience Methods, 203 (2012), pp. 200-211.

            J. Polzehl and K. Tabelow,
            \emph{Structural adaptive smoothing in diffusion tensor imaging:
            The R package dti}, Journal of Statistical Software, 31 (2009) pp. 1--24.

            K. Tabelow, J. Polzehl, V. Spokoiny and H.U. Voss.
            \emph{Diffusion Tensor Imaging: Structural adaptive smoothing},
            NeuroImage 39(4), 1763-1773 (2008).
}
\keyword{ package }
\seealso{
  \href{https://cran.r-project.org/package=fmri}{fmri}
  \href{https://cran.r-project.org/package=aws}{aws}
  \href{https://cran.r-project.org/package=oro.nifti}{oro.nifti}
}
\examples{
  \dontrun{demo(dti_art)}
  \dontrun{demo(mixtens_art)}
}
